+++ services/sms/include/sms_send_indexer.h
@@ -94,8 +94,8 @@ public:
     void SetLangId(uint8_t langId);
     uint16_t GetMsgId() const;
     void SetMsgId(uint16_t msgId);
-    uint16_t GetDataBaseId() const;
-    void SetDataBaseId(uint16_t msgId);
+    int32_t GetDataBaseId() const;
+    void SetDataBaseId(int32_t msgId);
     bool GetIsMmsApp() const;
     void SetIsMmsApp(bool isMmsApp);
 
@@ -129,7 +129,7 @@ private:
     bool isText_ = false;
     bool isMmsApp_ = true;
     uint16_t msgId_ = 0;
-    uint16_t dataBaseId_ = 0;
+    int32_t dataBaseId_ = 0;
 };
 } // namespace Telephony
 } // namespace OHOS
