+++ test/gtest/sms_mms_common_test.cpp
@@ -48,7 +48,7 @@ void SmsMmsCommonTest::TearDown() {}
 HWTEST_F(SmsMmsCommonTest, SendBroadcast_0001, Function | MediumTest | Level1)
 {
     std::string sendStatus = "";
-    uint16_t dataBaseId = 0;
+    int32_t dataBaseId = 0;
     auto smsCommon = DelayedSingleton<SmsMmsCommon>::GetInstance();
     smsCommon->SendBroadcast(dataBaseId, SmsMmsCommonData::SMS_MMS_SENT_RESULT_NOTIFY, sendStatus,
         SmsMmsCommonData::SMS_MMS_INFO_MMS_TYPE);
